In output/ref/ symlinks, strip the "origin/" prefix from branches.
Do this so the general public doesn't need to understand the funky
naming and quoting of "origin_", especially in their apt
sources.lists.

This can cause collisions between tags and branches, but we're just
going to ignore that for now.

When this is deployed on the binary tarball gitbuilders, the
corresponding URLs need to change in teuthology and ceph-autotests.