run-real: touch symlinks to failed builds
This way the symlink will stick around if the latest commit fails to build,
and we'll always have the last pass availble.

Keeping the symlink fresh means that the remote build result will also be
touched and kept around.

Having the latest passing build is better than nothing at all!