build-auto: determine gitbuilder behavior from the hostname
Figure out what the build should do based on keywords in the
hostname.  Similarly, rsync to the remote target using the
fields in the hostname:

 gitbuilder-PROJECT-WHAT-DISTRO-ARCH-FLAVOR

 PROJECT is ceph, ...
 WHAT is deb or rpm (so far)
 DISTRO is precise, etc.
 ARCH is amd64, i386 (although the VM knows this anyway)
 FLAVOR is basic, notcmalloc, gcov, ...

This will simplify the fabfile a bit (at least for new builder VMs).  We
also solve the problem is passing the environment into the package build
from one central location (build-auto.sh) which has enough context/state in
the hostname to do the right thing.

Signed-off-by: Sage Weil <sage@inktank.com>