build-ceph-rpm.sh:  Fix release candidate version
Version numbers for release candidate builds have -rc[1-9] suffix.
The '-' is an illegal character in rpm version strings. So, before
building, move the rc tag from the version string and add it to
the release string if it is not already there.  The rc tag does
need to be added to the source and setup lines since it is no longer
version.

Signed-off-by: Gary Lowell  <gary.lowell@inktank.com>