If git mirror file set use git mirror.
A bit of a hack but unfortunately with howe we are currently
removing the submodules to re-sync them with possible new URLs
it requires this to be somewhat dirty. Still doing the new URL
sync so atleast for non-mirror'd submodules it should still be
able to handle the URL changing.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>