Add support for building with Clang's static-analyzer.
Add support for wrapping the top-level 'make' invocation with the
contents of the BUILD_WRAPPER environment variable, and define this
variable on hosts with the -analyze suffix present.

Also disable the use of CCACHE in this case; it seems likely that
these two facilities might interact poorly.  This might be
overly conservative, however.

Note that while the static-analyzer is part of the clang project,
its use does not imply that compilation must use clang; hence if
you wish to both static-analyze and build Ceph with clang, then
both the -analyze and -clang hostname suffixes are required.