fabfile: always use gitbuilder master
Instead of a hardcoded SHA. If a new version of master at
https://github.com/ceph/gitbuilder/commits/master is available, it needs
to be used because it presumably contains a fix. If there is a need to
update the fabfile to take advantage of that fix, it's like having pull
request review twice on a commit. Once when it is reviewed prior to
being merged in gitbuilder and once when it is reviewed to be used in
the fabfile.

Signed-off-by: Loic Dachary <loic@dachary.org>