build-ceph-{deb-native,rpm}.sh: Clean up build directories on failure
A recent change to a wip branch was leaving a .git file (as in a submodule
pointer to ../../.git/modules/) behind in the out~ or rpmbuild/BUILD
dirs when failing the build, and then because these scripts both have -e,
they immediately exited.

Then, when the next build came around, it tried "git clean", which
recursed into the build dir, found the .git, tried to dereference into
the ultimate dir, and failed.

Harden against this sort of failure by making sure to remove the workdir
immediately, and then fail, when the actual build worker fails.

Signed-off-by: Dan Mick <dan.mick@redhat.com>